+++ src/view/router/router-index.js
@@ -3,12 +3,12 @@ import {
   HashRouter as Router,
   Route,
   // Switch,
-  Redirect
+  Redirect,
 } from "react-router-dom";
 
 import Loading from "../page/layout-loading";
 import Home from "../page/layout-home";
-import ChangeLog from "../page/layout-changelog";
+import Font from "../page/layout-font";
 import { AnimatedSwitch as Switch } from "react-router-transition";
 
 export default function RouterRelation() {
@@ -24,7 +24,8 @@ export default function RouterRelation() {
         <Route exact path="/" component={Loading} />
         <Route path="/loading" component={Loading} />
         <Route path="/home" component={Home} />
-        <Route path="/changelog" component={ChangeLog} />
+        <Route path="/font" component={Font} />
+        {/* <Route path="/changelog" component={ChangeLog} /> */}
         <Redirect to="/loading" />
       </Switch>
     </Router>